Fix #3188: Handle initial knockouts event correctly
It appears that I wasn't aware that the event source sends an initial
set of all its events when a page loads. This was results in the
corner pages doing an initial update based on the knockouts event
at load time. This in turn was forcing the use of the next upcoming
matches since that was all that it was getting there.

Modify the knockout event handling so that it only gets upcoming
matches if there are is no current match. If there is a current
match then re-fetch that instead.

Change-Id: Id5c7f7a0c72d1e9c5f13a3d6e21756844c611784