Adjust time columns to better serve the shepherds
This changes the Staging screen to show information relating to
the arenas and staging, in particular the deadlines they need to
know about. Here the 'current' match is the one which is currently
in staging.

The Shepherding screen now focusses on the match which the shepherds
are currently fetching, showing the countdowns to when they should
fetch teams (drive from signal_shepherds) and close the staging
area. Here the definition of the 'current' match varies depending
on whether the shepherds are signalled before or after the staging
opens and is either signal->opens or opens->closes.

Feedback from the shepherds is needed as to whether this is more
useful, and whether the staging open/close columns are on the right
screens.

Change-Id: If0964d6fc54466b91ffcb8d2e0cd31f891161e43