Ensure the screens highlight the right match
This changes to making the current time an explicit part of the
calculation of the class applied to the rows. In turn this means
that Polymer is able to detect changes in the class via changes
in the time. This ensures that the highlights are applied as soon
as the time rolls around, rather than needing to wait for the
schedule to be updated.

Change-Id: I1f75f5eadd71d3b22b47e165b064ffe030a2c97b