Rename 'matches' to avoid name conflict in Firefox
I don't understand why this works in Chrome but not Firefox (I do
suspect it's to do with whether it's using the polyfill or not),
but it appears that we can't have a property called matches on our
polymer things. Use 'matchList' as an alternative with a comment
about the reason for the less-than-ideal name.

Change-Id: Ib3f666e071842dd53b7aaa0e5ab6227cbb96875f