Cycle outlines when teams are touched on the shepherding screen
This is useful for keeping track of the status of teams who are (or
should be) on their way to the arena. As head shepherd, when one of my
shepherds has told a team that they're up soon, I touch their TLA to
outline it with dots. When they've been seen heading to the arena, I
touch it again for a dashed outline, and once they're seen in the Cube I
touch a final time to a show a solid border. This provides an
at-a-glance overview of where teams are, how the next match lineup is
looking, and who we need to chase.